fix: mettre à jour la version du plugin à 1.11.4 et corriger les liens de téléchargement

+++ EorzeaEventsPlugin/Plugin.cs
@@ -708,7 +708,7 @@ public sealed class Plugin : IDalamudPlugin
         return row?.PlaceName.Value.Name.ToString();
     }
 
-    private void OnTerritoryChanged(uint territory)
+    private void OnTerritoryChanged(ushort territory)
     {
         CurrentZone = ResolveTerritoryName(territory);
